indexer/requirements.txt
appdirs
colorlog
sha3
pycoin
python-bitcoinlib
python-bitcoinrpc
regex
pycryptodome
requests
cachetools
pymysql
python-docx
cryptography
pybase64
python-magic
boto3 # required if using AWS S3 for file storage
msgpack
#arweave-python-client # required for arweave storage